+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Document type does not allow empty available views ([https://github.com/GotCms/GotCms/issues/170](https://github.com/GotCms/GotCms/issues/170))
 - Translator with javascript have missing translations ([https://github.com/GotCms/GotCms/issues/165](https://github.com/GotCms/GotCms/issues/165))
 - Rebuild all forms ([https://github.com/GotCms/GotCms/issues/172](https://github.com/GotCms/GotCms/issues/172))
+- currentDocument must be an helper instead of variable ([https://github.com/GotCms/GotCms/issues/173](https://github.com/GotCms/GotCms/issues/173))
 
 ##1.1.0 (11 Dec 2013)
 - Update to Zend Framework 2.2.5 ([https://github.com/GotCms/GotCms/issues/141](https://github.com/GotCms/GotCms/issues/141))